requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pipreqs==0.4.13
platformdirs==3.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.3
yarg==0.1.9
